Moved local server servlets into net.lightbody.bmp.proxy.test subpackages
