+++ browsermob-core/src/test/java/net/lightbody/bmp/proxy/test/servlet/SetCookieServlet.java
@@ -1,4 +1,4 @@
-package net.lightbody.bmp.proxy;
+package net.lightbody.bmp.proxy.test.servlet;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.Cookie;
